+++ examples/using_arcs.py
@@ -3,7 +3,7 @@
 # Copyright (c) 2018 Manfred Moitzi
 # License: MIT License
 import ezdxf
-from ezdxf.algebra import Vector, Arc, UCS
+from ezdxf.algebra import Vector, ConstructionArc, UCS
 
 dwg = ezdxf.new('R2000')
 modelspace = dwg.modelspace()
@@ -25,32 +25,32 @@ end_point_ucs = ucs.from_wcs(end_point_wcs)
 def_point_ucs = Vector(0, 0)  # origin of UCS
 
 # create arc in the xy-plane of the UCS
-arc = Arc.from_3p(start_point_ucs, end_point_ucs, def_point_ucs)
+arc = ConstructionArc.from_3p(start_point_ucs, end_point_ucs, def_point_ucs)
 arc.add_to_layout(modelspace, ucs, dxfattribs={'color': 1})  # red arc
 
-arc = Arc.from_3p(end_point_ucs, start_point_ucs, def_point_ucs)
+arc = ConstructionArc.from_3p(end_point_ucs, start_point_ucs, def_point_ucs)
 arc.add_to_layout(modelspace, ucs, dxfattribs={'color': 2})  # yellow arc
 
 p1 = Vector(0, -18)
 p2 = Vector(0, +18)
-arc = Arc.from_2p_angle(p1, p2, 90)
+arc = ConstructionArc.from_2p_angle(p1, p2, 90)
 arc.add_to_layout(modelspace, dxfattribs={'color': 1})
 
-arc = Arc.from_2p_angle(p1, p2, 90, ccw=False)
+arc = ConstructionArc.from_2p_angle(p1, p2, 90, ccw=False)
 arc.add_to_layout(modelspace, dxfattribs={'color': 2})
 
 p1 = Vector(20, -18)
 p2 = Vector(20, +18)
-arc = Arc.from_2p_radius(p1, p2, 100)
+arc = ConstructionArc.from_2p_radius(p1, p2, 100)
 arc.add_to_layout(modelspace, dxfattribs={'color': 1})
 
-arc = Arc.from_2p_radius(p1, p2, 100, ccw=False)
+arc = ConstructionArc.from_2p_radius(p1, p2, 100, ccw=False)
 arc.add_to_layout(modelspace, dxfattribs={'color': 2})
 
-arc = Arc.from_2p_radius(p1, p2, 100, center_is_left=False)
+arc = ConstructionArc.from_2p_radius(p1, p2, 100, center_is_left=False)
 arc.add_to_layout(modelspace, dxfattribs={'color': 3})
 
-arc = Arc.from_2p_radius(p1, p2, 100, ccw=False, center_is_left=False)
+arc = ConstructionArc.from_2p_radius(p1, p2, 100, ccw=False, center_is_left=False)
 arc.add_to_layout(modelspace, dxfattribs={'color': 4})
 
 # saving DXF file